Moving tracer mdata processing to a sep src
Moving processing of tracer meta data to a separate source
(for maintainability)

Also explicitly naming the variables to indicate where we
use/expect MPI world rank
